Updated CoC
I think we may have been a bit too hasty with our previous code of conduct. Given that we are Notre Dame students and as Blessed Father Basil Moreau states: "We must not educate the mind at the expense of the heart." As a result, I have decided to include the Rule of St. Benedict as an addition to our code of conduct, to exemplify the ideals that Notre Dame has instilled in us.